+++ shared/docs/v0.0.1/changelog.md
@@ -8,6 +8,7 @@
 - Spinner component created with stories and tests 
 - Tag component created with tests and stories 
 - Sheet component created
+- Search component with stories and tests added
 
 ## 🐛 Fixes: